+++ packages/playwright/__tests__/knapsack-playwright.spec.ts
@@ -1,0 +1,30 @@
+import { describe, it, expect } from 'vitest';
+import { normalizePaths } from '../src/utils';
+
+describe('#normalize', () => {
+  it('concatenates the recorded paths assigning 0 seconds to the non-recorded ones', () => {
+    const paths = [
+      'x.spec.js',
+      'y.spec.js',
+      'z.spec.js',
+      'c.spec.js',
+      'd.spec.js',
+    ];
+
+    const recordedPaths = {
+      'c.spec.js': 4,
+      'd.spec.js': 3,
+    };
+
+    const actual = normalizePaths(paths, recordedPaths);
+    const expected = {
+      'x.spec.js': 0,
+      'y.spec.js': 0,
+      'z.spec.js': 0,
+      'c.spec.js': 4,
+      'd.spec.js': 3,
+    };
+
+    expect(actual).toEqual(expected);
+  });
+});